requirements.txt
streamlit
llama-index-core
llama-index-vector-stores-qdrant
llama-index-embeddings-huggingface
llama-index-llms-openai
qdrant-client
nest_asyncio
